fix: Refactor user menu
