+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.13.1] - 2023-03-14
+
+### Changed
+
+- Refactored user menu, adding a new `MenuEntry` component
+
 ## [1.13.0] - 2023-03-12
 
 ### Added
