+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bcc",
-  "version": "1.13.0",
+  "version": "1.13.1",
   "private": true,
   "scripts": {
     "dev": "next dev",
